# perf(core): avoid copying back to mirror when fetching from it